requirements.txt
alembic==1.12.1
asgiref==3.7.2
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.1.3
beautifulsoup4==4.11.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
dictdiffer==0.9.0
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
git-filter-repo==2.38.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigquery==3.22.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.21.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
grpcio-tools==1.62.2
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
install==1.3.5
ipdb==0.13.13
ipython==7.34.0
jedi==0.19.1
kthread==0.2.3
loguru==0.7.2
lxml==5.2.2
Mako==1.2.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
ncclient==0.6.9
numpy==1.21.6
packaging==24.0
pandas==1.3.5
paramiko==3.4.0
parso==0.8.4
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.24.4
psycopg2==2.9.9
psycopg2-binary==2.9.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygments==2.17.2
pygnmi==0.6.8
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.31
sqlparse==0.4.4
sshtunnel==0.4.0
tomli==2.0.1
traitlets==5.9.0
typing-extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
websocket-client==1.6.1
win32-setctime==1.1.0
xmltodict==0.13.0
zipp==3.15.0